Refactor GridFormInputGroup test assertions and renderers
